+++ consumer/arduino_serial_port.h
@@ -3,9 +3,9 @@
 
 #include <stdint.h>
 
-int arduino_serial_port_init(const char *serial_port, int baud_rate);
+int arduino_serial_port_init(const char* serial_port, size_t baud_rate);
 int arduino_serial_port_close(int fd);
 int arduino_serial_port_flush(int fd);
-int arduino_serial_port_read(int fd, unsigned char *buf, size_t buf_max, int sleep_micros);
+ssize_t arduino_serial_port_read(int fd, unsigned char* buf, size_t buf_max, unsigned int sleep_micros);
 
 #endif
